pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "octus-financial-agent"
version = "0.1.0"
description = "Financial Multi-Agent Intelligence System over Octus + SimFin data"
requires-python = ">=3.11"
dependencies = [
    # Core data
    "pandas>=2.2",
    "pyarrow>=16",
    "duckdb>=1.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "pyyaml>=6.0",
    # HTML parsing
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    # Embeddings / vector
    "sentence-transformers>=3.0",
    "pinecone>=4.1",
    "faiss-cpu>=1.8",
    "tiktoken>=0.7",
    # SimFin
    "simfin>=1.0",
    "httpx>=0.27",
    "rapidfuzz>=3.9",
    # LLM
    "anthropic>=0.28",
    # Frontend
    "streamlit>=1.35",
    "chainlit>=1.1",
    # Utilities
    "filelock>=3.14",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "responses>=0.25",
    "jupyter>=1.0",
    "ipykernel>=6.29",
    "matplotlib>=3.9",
    "plotly>=5.22",
    "nbconvert>=7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
